requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-filter>=24.3
django-mptt>=0.16
drf-spectacular>=0.27
python-decouple>=3.8
Pillow>=10.4
bcrypt>=4.2
filetype>=1.2
pywebpush>=2.0
channels>=4.1
daphne>=4.1
